[Refactor] Packaging ui components
