+++ src/main/java/com/example/pssupporter/ui/list/MyCellRenderer.java
@@ -2,7 +2,7 @@
  * Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package com.example.pssupporter.ui;
+package com.example.pssupporter.ui.list;
 
 import com.intellij.icons.AllIcons;
