+++ src/main/java/com/example/pssupporter/ui/list/MyTestListItem.java
@@ -2,8 +2,9 @@
  * Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package com.example.pssupporter.ui;
+package com.example.pssupporter.ui.list;
 
+import com.example.pssupporter.ui.editor.MyEditorPanel;
 import com.example.pssupporter.vo.TestStatus;
 
 public class MyTestListItem {
