+++ src/main/java/com/example/pssupporter/ui/list/MyTestListPanel.java
@@ -2,8 +2,9 @@
  * Copyright 2000-2023 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package com.example.pssupporter.ui;
+package com.example.pssupporter.ui.list;
 
+import com.example.pssupporter.ui.editor.MyEditorPanel;
 import com.example.pssupporter.utils.ComponentManager;
 import com.example.pssupporter.vo.TestData;
 import com.intellij.openapi.project.Project;
